move button, list and load_indicator from 'editors' and mode 'drop_down_editors' to 'editors'
