+++ js/renovation/ui/__tests__/button.test.tsx
@@ -1,26 +1,26 @@
 import React from 'react';
 import { mount, shallow } from 'enzyme';
-import devices from '../../../../core/devices';
-import { convertRulesToOptions } from '../../../../core/options/utils';
-import { current } from '../../../../ui/themes';
+import devices from '../../../core/devices';
+import { convertRulesToOptions } from '../../../core/options/utils';
+import { current } from '../../../ui/themes';
 import {
   clear as clearEventHandlers,
   defaultEvent,
   emit,
   getEventHandlers,
   EVENT,
-} from '../../../test_utils/events_mock';
+} from '../../test_utils/events_mock';
 import {
   Button, ButtonProps, defaultOptionRules, viewFunction,
 } from '../button';
-import { Widget } from '../../common/widget';
-import { Icon } from '../../common/icon';
-import { InkRipple } from '../../common/ink_ripple';
+import { Widget } from '../common/widget';
+import { Icon } from '../common/icon';
+import { InkRipple } from '../common/ink_ripple';
 
 interface Mock extends jest.Mock {}
 
-jest.mock('../../../../core/devices', () => {
-  const actualDevices = jest.requireActual('../../../../core/devices').default;
+jest.mock('../../../core/devices', () => {
+  const actualDevices = jest.requireActual('../../../core/devices').default;
   const isSimulator = actualDevices.isSimulator.bind(actualDevices);
   const real = actualDevices.real.bind(actualDevices);
 
@@ -30,8 +30,8 @@ jest.mock('../../../../core/devices', () => {
   return actualDevices;
 });
 
-jest.mock('../../../../ui/themes', () => ({
-  ...jest.requireActual('../../../../ui/themes'),
+jest.mock('../../../ui/themes', () => ({
+  ...jest.requireActual('../../../ui/themes'),
   current: jest.fn(() => 'generic'),
 }));
