+++ js/renovation/ui/editors/drop_down_editors/date_box.tsx
@@ -1,10 +1,10 @@
 import {
   Component, ComponentBindings, JSXComponent, OneWay, TwoWay, Event,
 } from '@devextreme-generator/declarations';
-import LegacyDateBox from '../../../ui/date_box';
-import { DomComponentWrapper } from '../common/dom_component_wrapper';
-import { EventCallback } from '../common/event_callback';
-import { BaseWidgetProps } from '../common/base_props';
+import LegacyDateBox from '../../../../ui/date_box';
+import { DomComponentWrapper } from '../../common/dom_component_wrapper';
+import { EventCallback } from '../../common/event_callback';
+import { BaseWidgetProps } from '../../common/base_props';
 
 export const viewFunction = ({
   props,
